requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
Flask-Session==0.8.0
openai-whisper==20231117
pydub==0.25.1
httpx==0.27.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
Flask-APScheduler==1.13.1 